refacto: cozy bar initialisation using preact and JSX :construction:
