+++ src/styles/bar.css
@@ -11,6 +11,10 @@
   font-size: 1rem;
 }
 
+[role=banner] .coz-bar-container {
+  display: flex;
+}
+
 @media (max-width: 48em) {
   [role=banner] {
     padding: 0 1em 0 0;
